Struct renderer::Renderer

pub struct Renderer { /* fields omitted */ }

The main renderer value.

Methods

impl Renderer

pub fn draw<T, F>(&mut self, cb: F) where
F: FnOnce(Arc<SwapchainImage<Window>>) -> Result<T>,
T: 'static + CommandBuffer + Send,

Draws the command buffer created on the given image, printing errors instead of returning them.

pub fn draw_inner<T, F>(&mut self, cb: F) -> Result<()> where
F: FnOnce(Arc<SwapchainImage<Window>>) -> Result<T>,
T: 'static + CommandBuffer + Send,

Draws the command buffer created on the given image.

impl Renderer

pub fn new(
name: Option<&str>,
version: Option<(u16, u16, u16)>
) -> Result<(Renderer, EventsLoop)>

Creates a new Renderer with the given application name and version.